td-shim-tools: merge td-shim-ld into td-shim-tools
Merge td-shim-ld into td-shim-tools.
