+++ .github/workflows/main.yml
@@ -61,21 +61,17 @@ jobs:
           AR: llvm-ar
         with:
           command: xbuild
-          args: -p td-shim --target x86_64-unknown-uefi --release --features=tdx
+          args: -p td-shim --target x86_64-unknown-uefi --release --features=main,tdx
 
       - name: Build PE format payload
         run: |
-          pushd td-payload
-          cargo xbuild --target x86_64-unknown-uefi --release --features=tdx
-          popd
-          cargo run -p td-shim-ld -- target/x86_64-unknown-uefi/release/ResetVector.bin target/x86_64-unknown-uefi/release/td-shim.efi target/x86_64-unknown-uefi/release/td-payload.efi -o target/x86_64-unknown-uefi/release/final.bin
+          cargo xbuild -p td-payload --target x86_64-unknown-uefi --release --features=main,tdx
+          cargo run -p td-shim-tools --bin td-shim-ld -- target/x86_64-unknown-uefi/release/ResetVector.bin target/x86_64-unknown-uefi/release/td-shim.efi target/x86_64-unknown-uefi/release/td-payload.efi -o target/x86_64-unknown-uefi/release/final-pe.bin
 
       - name: Build Elf format payload
         run: |
-          pushd td-payload
-          cargo xbuild --target ../devtools/rustc-targets/x86_64-unknown-none.json --release --features=tdx
-          popd
-          cargo run -p td-shim-ld -- target/x86_64-unknown-uefi/release/ResetVector.bin target/x86_64-unknown-uefi/release/td-shim.efi target/x86_64-unknown-none/release/td-payload -o target/x86_64-unknown-uefi/release/final.bin
+          cargo xbuild -p td-payload --target devtools/rustc-targets/x86_64-unknown-none.json --release --features=main,tdx
+          cargo run -p td-shim-tools --bin td-shim-ld -- target/x86_64-unknown-uefi/release/ResetVector.bin target/x86_64-unknown-uefi/release/td-shim.efi target/x86_64-unknown-none/release/td-payload -o target/x86_64-unknown-uefi/release/final-elf.bin
 
   test:
     name: Test
