+++ Cargo.toml
@@ -6,7 +6,6 @@ members = [
     "devtools/td-benchmark",
     "devtools/test-runner-server",
     "devtools/test-runner-client",
-    "td-shim",
     "fuzzing/fuzz_pe_loader",
     "fuzzing/fuzz_elf_loader",
     "fuzzing/fuzz_fv_parser",
@@ -16,8 +15,8 @@ members = [
     "td-logger",
     "td-paging",
     "td-payload",
+    "td-shim",
     "td-shim-tools",
-    "td-shim-ld",
     "tdx-tdcall",
     "tests/test-td-exception",
     "tests/test-td-paging",
