+++ td-payload/Cargo.toml
@@ -8,7 +8,6 @@ license = "BSD-2-Clause-Patent"
 edition = "2018"
 
 [dependencies]
-td-benchmark = { path = "../devtools/td-benchmark", optional = true }
 chrono = { version = "0.4.19", default-features = false, features = ["serde"]}
 linked_list_allocator = "0.9.0"
 log = "0.4.13"
@@ -19,6 +18,7 @@ td-layout = { path = "../td-layout" }
 td-logger =  { path = "../td-logger" }
 uefi-pi =  { path = "../uefi-pi" }
 
+td-benchmark = { path = "../devtools/td-benchmark", optional = true }
 tdx-tdcall = { path = "../tdx-tdcall", optional = true }
 
 [features]
@@ -26,3 +26,4 @@ default = []
 cet-ss = []
 benches = ["td-benchmark"]
 tdx = ["tdx-tdcall", "td-logger/tdx"]
+main = []
