--- td-shim-ld/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "td-shim-ld"
-version = "0.1.0"
-description = "TD shim linker to compose multiple components into shim binary"
-repository = "https://github.com/confidential-containers/td-shim"
-homepage = "https://github.com/confidential-containers"
-license = "BSD-2-Clause-Patent"
-edition = "2018"
-
-[features]
-default = ["linker"]
-boot-kernel = ["td-layout/boot-kernel"]
-linker = ["clap", "env_logger", "log", "pe-loader"]
-
-[dependencies]
-r-efi = "3.2.0"
-scroll = { version = "0.10", default-features = false }
-td-layout = { path = "../td-layout" }
-uefi-pi =  { path = "../uefi-pi" }
-
-clap = { version = "3.0", features = ["cargo"], optional = true }
-env_logger = { version = "0.9.0", optional = true }
-log = { version = "0.4.5", optional = true }
-pe-loader = { path = "../pe-loader", optional = true }
