+++ td-shim-tools/Cargo.toml
@@ -12,6 +12,10 @@ name = "td-shim-enroll-key"
 required-features = ["enroller"]
 
 [[bin]]
+name = "td-shim-ld"
+required-features = ["linker"]
+
+[[bin]]
 name = "td-shim-sign-payload"
 required-features = ["signer"]
 
@@ -20,16 +24,18 @@ r-efi = "3.2.0"
 scroll = { version = "0.10", default-features = false, features = ["derive"]}
 td-layout = { path = "../td-layout" }
 td-shim = { path = "../td-shim" }
-td-shim-ld = { path = "../td-shim-ld", default-features = false }
 uefi-pi =  { path = "../uefi-pi" }
 
 clap = { version = "3.0", features = ["cargo"], optional = true }
 der = { version = "0.4.5", features = ["oid"], optional = true }
 env_logger = { version = "0.9.0", optional = true }
 log = { version = "0.4.5", optional = true }
+pe-loader = { path = "../pe-loader", optional = true }
 ring = { git = "https://github.com/jyao1/ring.git", branch = "uefi_support", default-features = false, features = ["alloc"], optional = true }
 
 [features]
-default = ["enroller", "signer"]
-enroller = ["clap", "der", "env_logger", "log", "ring", "td-shim/secure-boot", "td-shim-ld/linker"]
-signer = ["clap", "der", "env_logger", "log", "ring", "td-shim/secure-boot", "td-shim-ld/linker"]
+default = ["boot-kernel", "enroller", "linker", "signer"]
+boot-kernel = ["td-layout/boot-kernel"]
+enroller = ["clap", "der", "env_logger", "log", "ring"]
+linker = ["clap", "env_logger", "log", "pe-loader"]
+signer = ["clap", "der", "env_logger", "log", "ring"]
