+++ td-shim/src/bin/td-shim/linux/boot.rs
@@ -2,19 +2,20 @@
 //
 // SPDX-License-Identifier: BSD-2-Clause-Patent
 
-use crate::e820::E820Entry;
-use crate::linux::kernel_param::{BootParams, SetupHeader};
 use core::mem::size_of;
-use rust_tdshim::{PayloadInfo, TdKernelInfoHobType};
 use scroll::{Pread, Pwrite};
 use td_layout as layout;
 use td_layout::runtime::TD_PAYLOAD_PARAM_BASE;
+use td_shim::{PayloadInfo, TdKernelInfoHobType};
 use x86_64::{
     instructions::{segmentation::Segment, tables::lgdt},
     registers::segmentation as seg,
     structures::{gdt, DescriptorTablePointer},
     PrivilegeLevel as RPL, VirtAddr,
 };
+
+use crate::e820::E820Entry;
+use crate::linux::kernel_param::{BootParams, SetupHeader};
 
 const KERNEL_64BIT_ENTRY_OFFSET: u64 = 0x200;
 const GDT: [u64; 4] = [
