+++ td-shim/src/bin/td-shim/linux/kernel_param.rs
@@ -2,9 +2,10 @@
 //
 // SPDX-License-Identifier: BSD-2-Clause-Patent
 
-use crate::e820::E820Entry;
 use core::mem::size_of;
 use zerocopy::{AsBytes, FromBytes};
+
+use crate::e820::E820Entry;
 
 #[derive(Clone, Copy, Default, Debug, AsBytes, FromBytes)]
 #[repr(C, packed)]
